requirements/base.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coveralls==1.3.0
decorator==4.4.1
Django==2.2.13
djangorestframework==3.10.3
django-filter==2.2.0
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
djangorestframework-filters
djangorestframework-jwt
djangorestframework-simplejwt
django-environ==0.4.5  # https://github.com/joke2k/django-environ
Markdown==3.1.1
docopt==0.6.2
idna==2.9
ipython-genutils==0.2.0
jedi==0.15.2
nose==1.3.7
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.5.2
python-decouple==3.1
pytz==2019.3
requests==2.23.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower